pyproject.toml
[project]
name = "habits-traker"
version = "0.1.0"
description = ""
authors = [
    {name = "Ananim4eg",email = "109145711+Ananim4eg@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
